fix props preview in docs
