+++ packages/circuit-ui/components/SidePanel/SidePanel.mdx
@@ -14,7 +14,8 @@ You can also open a secondary side panel on top of the first, triggered from an 
 _Note: the examples for the side panel component are best viewed in canvas mode_
 
 <Story of={Stories.Base} />
-<Props />
+<Props
+         include={['headline', 'children', 'onClose', 'closeButtonLabel',  'backButtonLabel', 'group']}  />
 
 ## Usage guidelines
